working network

+++ scripts/socket.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-rm sockets/$1.socket
+[ -e sockets/$1.socket ] && rm sockets/$1.socket
 
 bin/firecracker --api-sock sockets/$1.socket
